+++ Assignment2/Boatplane.cpp
@@ -14,13 +14,13 @@ namespace assignment2
 
 	unsigned int Boatplane::GetFlySpeed() const
 	{
-		int x = static_cast<int>(GetPassengersTotalWeight());
+		int x = GetPassengersTotalWeight();
 
 		return CalPlaneSpeed(x, 150, 500, 300.0);
 	}
 	unsigned int Boatplane::GetSailSpeed() const
 	{
-		int x = static_cast<int>(GetPassengersTotalWeight());
+		int x = GetPassengersTotalWeight();
 		return CalSailSpeed(x, 1.7);
 	}
 	unsigned int Boatplane::GetMaxSpeed() const
